### 2019-01-28 #1211 @whedon
[PRE REVIEW]: Detecting Fraud in Online Surveys by Tracing, Scoring, and Visualizing IP Addresses
**Submitting author:** @pdwaggoner (<a href="http://orcid.org/0000-0002-7825-7573">Philip D. Waggoner</a>)
**Repository:** <a href="https://github.com/MAHDLab/rIP" target ="_blank">https://github.com/MAHDLab/rIP</a>
**Version:** v1.1.1
