### 2019-01-28 #1211 @arfon
Message from the submitting author:

> Our paper details our recent R package, which detects fraud in online surveys by tracing, scoring, and visualizing IP addresses, after cleaning IP data. In our previous submission, @arfon (rightly) noted the package was in the "minor utility" category. As a response, we have worked to significantly expand the package functionality as well as including a variety of cleaning and plotting tools, all with the goal of making survey research more efficient and reproducible. Our package is very important for practical research that uses online surveys, from marketing to political science, and many other fields. Thank you for considering.
