+++ paper.bib
@@ -99,23 +99,4 @@
   volume   = {111},
   number   = {4},
   pages    = {493--504}
-  %abstract = {The authors find that experimental studies using online samples
-    %          (e.g., MTurk) often violate the assumption of random assignment,
-      %        because participant attrition-quitting a study before completing
-        %      it and getting paid-is not only prevalent, but also varies
-          %    systemically across experimental conditions. Using standard
-   %           social psychology paradigms (e.g., ego-depletion, construal
-     %         level), they observed attrition rates ranging from 30\% to 50\%
-       %       (Study 1). The authors show that failing to attend to attrition
-         %     rates in online panels has grave consequences. By introducing
-           %   experimental confounds, unattended attrition misled them to draw
-  %            mind-boggling yet false conclusions: that recalling a few happy
-    %          events is considerably more effortful than recalling many happy
-      %        events, and that imagining applying eyeliner leads to weight loss
-        %      (Study 2). In addition, attrition rate misled them to draw a
-          %    logical yet false conclusion: that explaining one's view on gun
-            %  rights decreases progun sentiment (Study 3). The authors offer a
-   %           partial remedy (Study 4) and call for minimizing and reporting
-     %         experimental attrition in studies conducted on the Web. (PsycINFO
-       %       Database Record},
 }
