+++ README.md
@@ -13,3 +13,5 @@ Especially important in this is the variable "block", which gives a score indica
 `rIP` requires users to have active (free) accounts and/or valid keys at iphub, ipintel, and/or proxycheck.
 
 We thank @tylerburleigh for his help on this tool. His method for incorporating this information into Qualtrics surveys can be found [here](https://twitter.com/tylerburleigh/status/1042528912511848448?s=19).
+
+See some additional working papers [here](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3272468) and [here](https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3327274). 
