### 2019-02-09 #1211 @yochannah
Hey @pdwaggoner! Before we progress to the review stage, JOSS requires that all repos have an OSI-approved license. Any of the licenses here should be fine: https://opensource.org/licenses

If this is okay with you, let me know when you've updated the license in your GitHub repo so we can proceed 👍
