### 2019-02-10 #1211 @pdwaggoner
Many thanks for editing @yochannah ! The license file should be in the repo and up to date, as LICENSE. Let me know if you aren’t seeing it. I’ll do whatever is needed. Thanks again!
