### 2019-02-10 #1211 @yochannah
Thanks @pdwaggoner - when I visit https://github.com/MAHDLab/rIP/blob/master/LICENSE I do see a license file, but it looks like it's just a couple of lines of copyright statement - [this is not technically enough to allow legal re-use of the software](https://choosealicense.com/no-permission/). OSI licenses have specific wording that explicitly allow re-use of the software, which is why we require them for JOSS.

A couple of resources that could help you choose which is right for you might be https://choosealicense.com/ and/or https://tldrlegal.com/. Personally, I'm a fan of MIT or Apache 2 licenses - they allow re-use relatively freely such that it can be re-used in either open source software or proprietary software, as opposed to a license like the GPL which only allows re-use by other open source software.
