### 2019-02-10 #1211 @pdwaggoner
Many thanks @yochannah - that makes good sense. I updated the license. Let me know if this is sufficient to move forward. The old version was our license file for our CRAN submission. Should be good now. Thanks!
