### 2019-02-25 #1211 @pdwaggoner
Sounds good - @yochannah . Thanks for taking that extra step; we really appreciate the added effort there. And Thanks as well @alexhanna ! Honestly, I don't think it will be too much of a taxing review, as the software is pretty straightforward, even if R is not your primary language. However, you both would like to proceed, we are more than happy to go along with it. Thanks again for the help! Excited to move forward on this.
