### 2019-02-26 #1211 @alexhanna
For sure. Let me know what the next steps are.

On Mon, Feb 25, 2019 at 2:29 PM Philip Waggoner <notifications@github.com>
wrote:

> Sounds good - @yochannah <https://github.com/yochannah> . Thanks for
> taking that extra step; we really appreciate the added effort there. And
> Thanks as well @alexhanna <https://github.com/alexhanna> ! Honestly, I
> don't think it will be too much of a taxing review, as the software is
> pretty straightforward, even if R is not your primary language. However,
> you both would like to proceed, we are more than happy to go along with it.
> Thanks again for the help! Excited to move forward on this.
>
