### 2019-02-27 #1285 @whedon
[REVIEW]: Detecting Fraud in Online Surveys by Tracing, Scoring, and Visualizing IP Addresses
**Submitting author:** @pdwaggoner (<a href="http://orcid.org/0000-0002-7825-7573">Philip D. Waggoner</a>)
**Repository:** <a href="https://github.com/MAHDLab/rIP" target ="_blank">https://github.com/MAHDLab/rIP</a>
**Version:** 1.1.2
