### 2019-04-19 #1285 @alexhanna
Hi @pdwaggoner -- I'm still reviewing it, sorry. Will try to have this
completed by tomorrow.

On Fri, Apr 19, 2019 at 9:02 AM Philip Waggoner <notifications@github.com>
wrote:

> hi @yochannah <https://github.com/yochannah>! Just wanted to follow up on
> this paper and see what the status is? No rush; just checking. Thanks!
>
