### 2019-04-21 #1285 @alexhanna
Hi @pdwaggoner,

This is almost there. There's a few pieces which I think would strengthen this submission.

- Could have an automated test or an example IP address. I'd like to have this to test particular IP addresses
- Needs to have some kind of contributor statement or CONTRIBUTING file
- Needs references DOIs
- It'd be helpful to have example usage and installation on the repo page
