### 2019-04-22 5e1c3d8
Update rIP.R
