+++ paper.bib
@@ -4,7 +4,7 @@
   journal={Research \& Politics},
   volume={2},
   number={3},
-  pages={2053168015604648},
+  doi={10.1177/2053168015604648},
   year=2015
 }
 
@@ -14,7 +14,7 @@
   journal={Research \& Politics},
   volume={2},
   number={4},
-  pages={2053168015622072},
+  doi={10.1177/2053168015622072},
   year=2015
 }
 
@@ -30,7 +30,8 @@
   title={Mturk Workers' Use of Low-Cost "Virtual Private Servers" to Circumvent Screening Methods: A Research Note},
   author={Dennis, Sean A and Goodson, Brian M and Pearson, Chris},
   year=2018,
-  institution = {Working Paper: University of Kentucky}
+  doi={10.2139/ssrn.3233954},
+  institution={Working Paper: University of Kentucky}
 }
 
 @online{team:2000,
@@ -48,6 +49,7 @@
   volume={29},
   number={6},
   pages={2156--2160},
+  doi={10.1016/j.chb.2013.05.009},
   year=2013
 }
 
@@ -58,6 +60,7 @@
   volume={6},
   number={1},
   pages={3--5},
+  doi={10.1177/1745691610393980},
   year=2011
 }
 
@@ -80,6 +83,7 @@
   title={The Shape of and Solutions to the MTurk Quality Crisis},
   author={Kennedy, Ryan and Clifford, Scott and Burleigh, Tyler and Jewell, Ryan and Waggoner, Philip D},
   year=2018,
+  doi={10.2139/ssrn.3272468},
   institution = {Working Paper: MAHD Lab, University of Houston}
 }
 
@@ -88,15 +92,17 @@
   title={Psychologists grow increasingly dependent on online research subjects},
   booktitle={Science: AAAS},
   year=2016,
+  doi={10.1126/science.aag0592},
   url={https://www.sciencemag.org/news/2016/06/psychologists-grow-increasingly-dependent-online-research-subjects}
 }
 
 @Article{Zhou:2016,
-  author   = {Zhou, Haotian and Fishbach, Ayelet},
-  title    = {The pitfall of experimenting on the web: How unattended selective attrition leads to surprising (yet false) research conclusions},
-  journal  = {J. Pers. Soc. Psychol.},
-  year     = 2016,
-  volume   = {111},
-  number   = {4},
-  pages    = {493--504}
+  author={Zhou, Haotian and Fishbach, Ayelet},
+  title={The pitfall of experimenting on the web: How unattended selective attrition leads to surprising (yet false) research conclusions},
+  journal={J. Pers. Soc. Psychol.},
+  year=2016,
+  volume={111},
+  number={4},
+  doi={10.1037/pspa0000056},
+  pages={493--504}
 }
