### 2019-04-22 #1285 @pdwaggoner
Hi @alexhanna (CC: @yochannah ) - Many thanks for the review! All great points to consider. I have updated everything accordingly:

1. I updated the example in the source code (the `rIP` file in the `R` folder in the repo) to include two real IPs, along with a real IP Hub key that any user could run successfully.

2. I added a `CONTRIBUTING` .txt file in the repo with corresponding responsibilities for each main author.

3. Added reference DOIs for all citations that had them. The following did not have DOIs:

- https://www.wired.com/story/amazon-mechanical-turk-bot-panic/
- https://cran.r-project.org/doc/manuals/r-release/R-lang.html
- https://blog.turkprime.com/after-the-bot-scare-understanding-whats-been-happening-with-data-collection-on-mturk-and-how-to-stop-it
- Ahler et al.: http://www.gsood.com/research/papers/turk.pdf

4. Finally, I included the same real example from the source code mentioned above in `#1` on the main repo page/README, along with installation instructions per your request.

5. Finally, I updated the acknowledgements in the main paper (`paper.md`) to thank both you and @yochannah  for your review of our paper and package.

Please don't hesitate to let me know anything else that may require attention as well as any next steps you'd like us to take. Thanks again!
