+++ R/rIP.R
@@ -11,41 +11,45 @@
 #'@details Takes an array of IPs and the keys for the services the user wishes to use (IP Hub, IP Intel, and Proxycheck), and passes these to all respective APIs. Returns a dataframe with the IP addresses (used for merging), country, ISP, labels for non-US IP Addresses, VPS use, and recommendations for blocking. The function also provides optional visualization tools for checking the distributions.
 #'@return ipDF A dataframe with the IP address, country code, country name, asn, isp, block, and hostname.
 #'@note Users must have active accounts and/or valid keys at iphub, ipintel, and/or proxycheck.
-#'@example
-#'library(rIP)
-#'ip_hub_key <- "MzI2MTpkOVpld3pZTVg1VmdTV3ZPenpzMmhodkJmdEpIMkRMZQ=="
-#'ipsample <- data.frame(rbind(c(1, "129.7.105.146"), c(2, "128.239.134.248")))
+#'@examples
+#'\dontrun{
+#'ip_hub_key <- "MzI2MTpkOVpld3pZTVg1VmdTV3ZPenpzMmhopIMkRMZQ=="
+#'ipintel_key <- "useremailaddress"
+#'proxycheck_key <- "MzI2MTpkOVpld3pZTVg1VmdTV3ZPenpzMmhod"
+#'ipsample <- data.frame(rbind(c(1, "189.8.105.146"), c(2, "148.233.134.248")))
 #'names(ipsample) <- c("number", "IPAddress")
-#'getIPinfo(ipsample, "IPAddress", iphub_key = ip_hub_key)
+#'getIPinfo(ipsample, "IPAddress", ip_hub_key, ipintel_key, proxycheck_key)
+#'}
 #'@export
-getIPinfo <- function(d, i, iphub_key = "", ipintel_key = "", proxycheck_key = "", plots = TRUE){
-  if (!requireNamespace("httr", quietly = TRUE)) {
+getIPinfo <- function(d, i, iphub_key, ipintel_key, proxycheck_key, plots = TRUE){
+  if(!requireNamespace("httr", quietly = TRUE)) {
     stop("Package \"httr\" needed for this function to work. Please install it.",
          call. = FALSE)
   }
-  if (!requireNamespace("utils", quietly = TRUE)) {
+  if(!requireNamespace("utils", quietly = TRUE)) {
     stop("Package \"utils\" needed for this function to work. Please install it.",
          call. = FALSE)
   }
-  if (!requireNamespace("iptools", quietly = TRUE)) {
+  if(!requireNamespace("iptools", quietly = TRUE)) {
     stop("Package \"iptools\" needed for this function to work. Please install it.",
          call. = FALSE)
   }
-  if (!requireNamespace("dplyr", quietly = TRUE)) {
+  if(!requireNamespace("dplyr", quietly = TRUE)) {
     stop("Package \"dplyr\" needed for this function to work. Please install it.",
          call. = FALSE)
   }
-  d <- as.data.frame(d)
+  #message("* Consider storing the ipDF as an object to write as an external df, e.g., write.csv(ipDF, 'ipDF.csv')")
   if (!is.data.frame(d)) warning("d must be a 'data.frame' object: Try 'as.data.frame(d)'")
   ips <- unique(d[ , i])
   options(stringsAsFactors = FALSE)
   iphub_url <- "http://v2.api.iphub.info/ip/"
   ipintel_url <- "http://check.getipintel.net/check.php?ip="
   proxycheck_url <- "http://proxycheck.io/v2/"
-  # Check iphub.info!= ""
[…]
