### 2019-04-23 53cf1e9
Update getIPinfo.Rd
