### 2019-05-07 67d95f2
Delete rIP.R
