### 2019-05-07 59dfa94
Delete DESCRIPTION
