+++ DESCRIPTION
@@ -1,23 +1,0 @@
-Package: rIP
-Type: Package
-Title: Detects Fraud in Online Surveys by Tracing, Scoring, and Visualizing IP Addresses
-Version: 1.1.1
-BugReports: https://github.com/MAHDLab/rIP/issues
-Authors@R: c(person("Ryan", "Kennedy", role = c("aut", "cre"), email = "rkennedy@uh.edu"),
-             person("Philip", "Waggoner", role = "aut", email = "philip.waggoner@gmail.com"),
-             person("Scott", "Clifford", role = "ctb"))
-Maintainer: Ryan Kennedy <rkennedy@uh.edu>
-Description: Takes an array of IPs and the keys for the services the user wishes to use (IP Hub, IP Intel, and Proxycheck), 
-    and passes these to all respective APIs. Returns a dataframe with the IP addresses (used for merging), 
-    country, ISP, labels for non-US IP Addresses, VPS use, and recommendations for blocking. The package 
-    also provides optional visualization tools for checking the distributions.
-Imports: httr, utils, iptools, dplyr, graphics
-License: MIT + file LICENSE
-Encoding: UTF-8
-LazyData: true
-RoxygenNote: 6.1.1
-NeedsCompilation: no
-Packaged: 2018-12-23
-Author: Ryan Kennedy [aut, cre],
-  Philip Waggoner [aut],
-  Scott Clifford [ctb]
