### 2019-05-12 #1285 @alexhanna
Sorry about the delay, @pdwaggoner. Thanks for the changes in the README.

I'm getting this error when I try to test.

```
> getIPinfo(ipsample, "IPAddress", iphub_key = ip_hub_key)
  |======================================================================| 100%Getting IP Hub information.
Error in getIPinfo(ipsample, "IPAddress", iphub_key = ip_hub_key) :
  argument "ipintel_key" is missing, with no default
```
