### 2019-05-13 f382324
Update rIP.R
