### 2019-05-13 #1285 @pdwaggoner
No worries @alexhanna . So we left that in there to alert users that they only supplied one (or two) of the three possible ip check keys. However, this shouldn't stop the function from still pinging the provided service based on the information provided. So for the error you provided, it still worked and should have produced a plot of two "clean" IPs (if you are replicating the example from the repo README). But we decided to still throw an error message (though not a stop error) to let the user know they could've supplied additional keys to check additional services. But it still works as intended with any or all of the keys provided.

Still, to address your concern, which highlights a point of confusion in the architecture of the package, I added this line to line 12 in the README:

> *Note*: `rIP` requires users to have active (free) accounts and/or valid keys at iphub, ipintel, and/or proxycheck. If users opt to use only one or not all three of these keys, then the function will throw a message letting you know as much. However, it will still work properly and check the supplied IPs based on any combination of keys the user supplies.

I hope this answers your question. Thanks!
