### 2019-05-14 #1285 @alexhanna
I see now. Would it be preferable to make this a warning rather than an error?

Also, I still believe you need to add a version release, per the reviewer guidelines (Version: Does the release version given match the GitHub release (v1.1.1)?)
