### 2019-05-14 #1285 @pdwaggoner
Thanks! We opted for the error for a couple reasons: first, it is slightly more efficient from our perspective, as we didn't "hard code" any error or warnings messages. It's just a default due to lack of supplied arguments, making the function run a bit quicker. And second, we like the idea of an error over a warning (but still not a stop error) as its a bit of a stronger flag to the user, "strongly suggesting" they take advantage of the other IP services supported by the function. We felt that an error message, while still allowing the function to work, would be more likely to get their attention to corroborate the findings from using less than the full 3 IP service options.

Also, thanks for the note! We updated minor things since the original submission (now 1.1.2). I just released the stable version here: https://github.com/MAHDLab/rIP/releases/tag/1.1.2 .

Anything else you need form us? Thanks again!
