### 2019-05-22 #1285 @danielskatz
As an Associate Editor-in-Chief, I'll take over from @yochannah who I know is very busy, since it looks like this is ready to go.

@pdwaggoner - please archive the current version of the software in zenodo or similar, and let me know the DOI of the archive
