### 2019-05-22 #1285 @pdwaggoner
Hi @danielskatz - thanks for the note and this sounds great! The software is archived at Zenodo, and the DOI is: 10.5281/zenodo.3178394 . Let me know whatever else you need. Thanks again!
