### 2019-05-23 56cf409
fixes to bib entries and some other small things
