### 2019-05-23 #1285 @danielskatz
👋 @pdwaggoner - there are a bunch of little issues in the paper, and I believe https://github.com/MAHDLab/rIP/pull/5 will fix them.  Please take a look and merge this, or otherwise resolve the broken references, figure placement, and other small issues.
