### 2019-05-23 #1285 @pdwaggoner
@danielskatz merged. Thanks for the fixes! Happy to take a look at another proof if you think it’s necessary.
