### 2019-05-23 #1285 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
